bonus-hugging-face-spaces/01-image-color-mood-analyzer/requirements.txt
gradio>=6.0
Pillow
scikit-learn
numpy
